# src/spa/src/dto/models/checkout-for-list-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface CheckoutForListDto {
  cardNumber?: null | string;
  checkoutDate?: string;
  dateReturned?: null | string;
  dueDate?: string;
  id?: number;
  status?: null | string;
  title?: null | string;
}