# src/spa/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Library Management System",
    "version": "V1"
  },
  "paths": {
...
    "/api/Checkout": {
      "put": {
        "tags": [
          "Checkout"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutForCheckInDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutForCheckInDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutForCheckInDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutForCheckInDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Checkout"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCheckoutDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCheckoutDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCheckoutDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BasketForCheckoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Checkout"
        ],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Checkout/{id}": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "operationId": "GetCheckout",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForDetailedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForDetailedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForDetailedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Checkout/asset/{libraryAssetId}": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "parameters": [
          {
            "name": "libraryAssetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Checkout/card/{libraryCardId}": {
      "get": {
        "tags": [
          "Checkout"
        ],
        "parameters": [
          {
            "name": "libraryCardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "OrderBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SearchString",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutForListDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddressDto": {
        "required": [
          "city",
          "stateId",
          "street",
          "zipcode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "stateId": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "$ref": "#/components/schemas/StateDto"
          },
          "zipcode": {
            "maxLength": 5,
            "minLength": 5,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "BasketForCheckoutDto": {
        "required": [
          "assets"
        ],
        "type": "object",
        "properties": {
          "libraryCardId": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "integer",
            "format": "int32"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibraryAssetForBasket"
            }
          }
        },
        "additionalProperties": false
      },
...
      "CheckoutForCheckInDto": {
        "type": "object",
        "properties": {
          "checkoutId": {
            "type": "integer",
            "format": "int32"
          },
          "isRenew": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CheckoutForDetailedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "libraryAsset": {
            "$ref": "#/components/schemas/LibraryAssetForListDto"
          },
          "libraryCard": {
            "$ref": "#/components/schemas/LibraryCardForDetailedDto"
          },
          "checkoutDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "dateReturned": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CheckoutForListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "checkoutDate": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "dateReturned": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetForBasket": {
        "type": "object",
        "properties": {
          "libraryAssetId": {
            "minimum": 0,
            "exclusiveMinimum": true,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "LibraryAssetForListDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "assetType": {
            "type": "string",
            "nullable": true
          },
          "authorName": {
            "type": "string",
            "nullable": true
          },
          "copiesAvailable": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "LibraryCardForDetailedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "cardNumber": {
            "type": "string",
            "nullable": true
          },
          "fees": {
            "type": "number",
            "format": "double"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "addressId": {
            "type": "integer",
            "format": "int32"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "libraryCardPhoto": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "photoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "StateDto": {
        "required": [
          "abbreviations",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "abbreviations": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}